various changes and tests

+++ admin/posts.php
@@ -27,20 +27,31 @@ $i=0;
     <?php include('templates/admin-nav.php'); ?>
     <div class="admin-dashboard-wrapper p-30 flex flex-column">
         <?php include('templates/admin-logout.php'); ?> 
-        <div class="admin-posts-btn flex gap-30 mv-50">
-            <a href="add-post.php" class="btn green-btn">ADD POST</a>
-            <a href="edit-post.php" class="btn yellow-btn">EDIT POST</a>
-            <a href="delete-post.php" class="btn red-btn">DELETE POST</a>
-        </div>
+        <?php if (!empty($posts)) { ?>
+            <div class="admin-posts-btn flex gap-30 mv-50">
+                <a href="add-post.php" class="btn green-btn">ADD POST</a>
+                <a href="edit-post.php" class="btn yellow-btn">EDIT POST</a>
+                <a href="delete-post.php" class="btn red-btn">DELETE POST</a>
+            </div>
+        <?php } ?>
         <div class="admin-posts-bottom mv-50">
             <div class="dashboard-posts-wrapper">
-                <div class="post-list-grid">
-                    <div class="grid-head grid mb-20 ph-10 c-light-grey">
-                        <p>ID</p>          
-                        <p>Title</p>          
-                        <p>Date</p>
-                    </div>
-                    <?php foreach ($posts as $post) { ?> 
+                <?php if (!empty($posts)) { ?>
+                    <div class="post-list-grid">
+                        <div class="grid-head grid mb-20 ph-10 c-light-grey">
+                            <p>ID</p>          
+                            <p>Title</p>          
+                            <p>Date</p>
+                        </div>
+                <?php } ?>
+                    <?php if (empty($posts)) { ?>
+                        <div class="flex gap-10 align-items-end c-light-grey">
+                            <p>So empty.</p>
+                            <a href="add-post.php" class="btn green-btn">ADD POST</a>
+                            <p>your first post.</p>
+                        </div>
+                    <?php } ?>
+                    <?php foreach ($posts as $post) { ?>
                         <div class="grid-body grid c-light-grey pv-10 ph-10 <?php if ( $i % 2 !=0 ) { echo "bg-senary"; } ?> ">
                             <div class="grid-body-id flex align-items-center gap-10">
                                 <input type="checkbox" name="selected-post" value="<?php echo $post['post_id'] ?>">
@@ -48,7 +59,7 @@ $i=0;
                             </div>   
                             <p><?php echo $post['post_title'] ?></p>     
                             <p><?php echo date('F j, Y ', $post['post_timestamp']); ?></p>    
-                    </div>      
+                        </div>  
                     <?php $i++; } ?>
                 </div>
             </div>
